RED-377 move rpc-config
