+++ packages/htlc/test/lib/rpc-config.ts
