+++ packages/redshift-utils/src/rpc/index.ts
@@ -1,4 +1,3 @@
 export * from './base-rpc-client';
 export * from './ethereum';
 export * from './utxo';
-export * from './rpc-config';
